prov/cxi: Add SW aligned rdzv to check for minimum length
When issuing a software get for a rendezvous handled by
software (HW could not process or get_en=0); protect
against minimum eager data. The belief is this should not
be required since we force rendezvous eager data to be
larger than cache-line size, but makes the code more
robust.

NETCASSINI-3635

Signed-off-by: Steve Welch <welch@hpe.com>

prov/cxi: Fix for FC during NIC hw-to-sw transition

When the NIC has initiated a hardware to software transition,
properly handle any flow control event (EQ Full, or Request Full)
that may occur before the software transition/on-loading of
unexpected messages is complete. The PtlTE will be re-enabled
in software managed mode when the transition completes and
flow control drops are processed.

NETCASSINI-3648

Signed-off-by: Steve Welch <welch@hpe.com>