Fix ctrl regression test failure on real hardware.
NETCASSINI-3670

Issue was a bad test design that exploited (incorrectly) the
coincidence that under NETSIM, the simulated NIC address of 0 is the
same as the NETSIM "real" address. On real hardware, the NIC address is
not 0, and tests failed.

Fixed to distinguish clearly between simulated and real addresses, even
if they happen to be numerically equivalent.

Signed-off-by: Joe Nemeth <joseph.nemeth@hpe.com>