prov/cxi: Handle C_EVENT_SEND events
Even though send events are disabled, hardware can generate a send event
in addition to the normal put/get/amo event when an error occurs. Prior
to this, a sent event would cause an abort to be tripped. This is
incorrect. The correct action is to just drop the send event and wait
for the normal event.

NETCASSINI-3774

Signed-off-by: Ian Ziemba <ian.ziemba@hpe.com>

prov/cxi: Verify user key is within CXI prov range

The CXI provider currently supports 2 bytes of keys These checks were
missing.

NETCASSINI-3773

Signed-off-by: Ian Ziemba <ian.ziemba@hpe.com>

prov/cxi: Handle reads and invalid std MR keys

A read to an invalid standard MR key will cause a target event to be
generated. It is safe to just ignore this event.

NETCASSINI-3773

Signed-off-by: Ian Ziemba <ian.ziemba@hpe.com>

prov/cxi: Increase MR key size to 4 bytes

This is to support applications which may have been using an RKEY
outside the 2 byte range when the CXI provider was not
enforcing/checking the RKEY range.

NETCASSINI-3773

Signed-off-by: Ian Ziemba <ian.ziemba@hpe.com>