prov/cxi: Extend rendezvous ID from 8 to 15 bits
Increase the total number of rendezvous operations that
can be in flight from 256 to 32K. The actual number that
can be in flight (which includes all unexpected rendezvous
that have not be matched at the target) is restricted by the
size of info->tx_attr->size.

There are two pools of IDs, 0 to 255 for FI_MSG sends and
256 to 32K-1 for FI_TAGGED send.

This lessens the potential for poorly designed applications
to deadlock due at least two ranks exhausting the limit
of outstanding rendezvous operations without posting a
matching receive (i.e. All are unexpected and will not
complete).

Active messaging in MPI can use FI_MULTI_RECV with FI_MSG,
however, this is internal and with relatively small number
of messages that should not deadlock.

NETCASSINI-4664

Signed-off-by: Steve Welch <welch@hpe.com>

prov/cxi: Add unit test for expanded RDZV ID space

For FI_TAGGED messaging add test to validate unexpected
messaging will not deadlock at 256 rendezvous messages.

NETCASSINI-4664

Signed-off-by: Steve Welch <welch@hpe.com>

prov/cxi: Document rendezvous messaging constraints

Document the FI_TAGGED and FI_MSG rendezvous messaging
constraints with regard to the number of rendezvous
operations that can be in-flight.

NETCASSINI-4664

Signed-off-by: Steve Welch <welch@hpe.com>