core: Undo temporary inc/dec of use_cnt
The temporary inc/dec of use_cnt was for a memhooks + CUDA is valid
check workaround. Since memhooks + CUDA has other deadlock issues,
keeping this temporary inc/dec is not needed. This temporary inc/dec was
causing the following issue.

Newly allocated MR cache entries start with use_cnt == 1 and the
list_entry not initialized. The expectation is that when use_cnt == 0,
the MR cache entry is moved to the LRU list thus setting the list_entry
to valid values. util_mr_entry_valid() decrements use_cnt but does not
add entry to LRU list. This can lead to segfaults in a threaded env.

NETCASSINI-5328

Signed-off-by: Ian Ziemba <ian.ziemba@hpe.com>