prov/cxi: Use -FI_ECONNREFUSED in zbcoll
NETCASSINI-5353

Use -FI_ECONNREFUSED instead of -FI_EINVAL in zbcoll when attempting to
join an endpoint to a group to which it does not belong, to match usage
in collectives.

Signed-off-by: Joe Nemeth <joseph.nemeth@hpe.com>

prov/cxi: Add 'quiet' flag to frmwk_init.

NETCASSINI-5338

Add quiet flag to allow frmwk_log0 to be cleanly used in a help message
for multnode tests.

Cleanup, and remove support for unused ENVIRONMENT variables.

Signed-off-by: Joe Nemeth <joseph.nemeth@hpe.com>

prov/cxi: Add TRACE functionality to libfabric

NETCASSINI-5380

Move the TRACE functionality, previously existing in the test framework
code, into a new cxip_trace.c module that is part of the CXI provider
source, to make the TRACE features available in test platforms we do not
control, such as MPI, or customer applications.

Signed-off-by: Joe Nemeth <joseph.nemeth@hpe.com>