prov/cxi: Change macro names.
NETCASSINI-5502

CXIP_COLL_MIN_FREE changed to CXIP_COLL_MIN_MULTI_RECV to avoid
confusion. In the collectives implementation, this value sets the
minimum amount of free space that must remain in a PTE buffer before
triggering a hardware-based unlinking of that buffer and a move to the
next buffer. This is associated with the CXI MULTI_RECV behavior, and
the name has been changed to emphasize that association.

CXIP_COLL_MAX_TX_SIZE has nothing to do with the transmission size,
which is a fixed 49 bytes for collectives, but instead has to do wih the
data payload size of 32 bytes.

Signed-off-by: Joe Nemeth <joseph.nemeth@hpe.com>

prov/cxi: Set RX discard soon enough.

NETCASSINI-5502

The implementation of the RX discard, which is required for applications
like MPI was taking its setting from the multicast object defined in the
fi_join_collective() operation, but this call is reached too late in the
initialization and can leak RX events before they are shut off, which can
break MPI applications.

The RX discard should always be enabled, EXCEPT for NETSIM regression
testing, where it must be disabled to allow the test code to count received
packets. Since this is a runtime environment feature, it should be set
using the is_netsim() function.

Signed-off-by: Joe Nemeth <joseph.nemeth@hpe.com>

prov/cxi: Set RX discard soon enough (fixed)

NETCASSINI-5507

Previous fix of this did not work if coll regressions were run on code
compiled with dev_platform != CXI_PLATFORM_NETSIM.

In general, RX discard should be true. The case where it needs to be
FALSE is not associated with the NETSIM flag, but with a specific set of
tests in coll.c, the --filter="coll_put*/*") tests. This requires that
this test be able to disable RX discard for the test, regardless of the
dev_platform setting.

This is done by adding a flag to the COMM_KEY_RANK object in the av_set
to indicate the proper setting of the discard flag in collective RX
events. This object passes through the entire stack, and can be used at
the appropriate time in setting up the PTE for collectives.

Signed-off-by: Joe Nemeth <joseph.nemeth@hpe.com>

prov/cxi: Reset timeout even if send fails.

NETCASSINI-5502

Any root transmission failure would result in an immediate retry, which
could saturate the TX and cause more problems than it solves.

Signed-off-by: Joe Nemeth <joseph.nemeth@hpe.com>

prov/cxi: remove coll_data->red_max.

NETCASSINI-5502

Passing this value through a chain of variables is confusing and prone
to error. The red_max value is a fixed property of each mc_obj, and should
be referenced directly.

Signed-off-by: Joe Nemeth <joseph.nemeth@hpe.com>

prov/cxi: Initialize reduction after use.

NETCASSINI-5502

Race conditions can cause the first use of a reduction structure to be
either from an initiating event on the node, or an incoming packet. This
implies that every reduction packet must be in a fully-initialized
(empty) state at startup and after reduction completion, and must take
all initialization data from the first source. Thus initialization must
be part of the cleanup pass at the end of a reduction.

Signed-off-by: Joe Nemeth <joseph.nemeth@hpe.com>

prov/cxi: Fix root race and deadlock.

NETCASSINI-5502

The root state machine makes an implicit assumption that the root will
not be the LAST participant in a reduction. This is simply not the case,
and if the root is the last participant, the reduction will hang.

Signed-off-by: Joe Nemeth <joseph.nemeth@hpe.com>

prov/cxi: Use coll.min_multi_recv.

NETCASSINI-5502

Code incorrectly used rxc.min_multi_recv, which could use different
values than specified for collectives.

Signed-off-by: Joe Nemeth <joseph.nemeth@hpe.com>

prov/cxi: Add buffer low-water counter.

NETCASSINI-5502

This is added for future diagnostic purposes.

Signed-off-by: Joe Nemeth <joseph.nemeth@hpe.com>

prov/cxi: Initialize red_cnt in _init_coll_data().

NETCASSINI-5502

The local coll_data structure red_cnt is initialized either by local
injection (which assigns a value of 1), or by receipt of a reduction
packet from the fabric, which can have higher values.

There are two places the value of 1 can be assigned, either on
injection, or on the first use of a software pre-reduction. The
pre-reduction case was missed. In both cases, _init_coll_data() is used
to initialize the local coll_data structure, so the assignment of 1 to
the count is moved to _init_coll_data().

Signed-off-by: Joe Nemeth <joseph.nemeth@hpe.com>

prov/cxi: Fix fi_query_collective() for fi_broadcast().

NETCASSINI-5549

Fix fi_query_collective().

Ignore the operator and datatype for FI_BARRIER.
Ignore the operator for FI_BROADCAST.

Signed-off-by: Joe Nemeth <joseph.nemeth@hpe.com>