prov/cxi: Fix status check for fetching AMO with flush
AMO with FI_DELIVERY_COMPLETE require two commands, the
AMO command status should be returned. The existing code
checks for not equal to C_EVENT_PUT type, but the status
for the flush is an EVENT_ACK. Keep the status for the
fetching AMO (C_EVENT_REPLY) and return that. This is exposed
if we do not abort the target of an AMO when a MR bounds error
occurs. In the bounds error case, we should report the error
to the initiator, just as we do for a bad MR key. A bad remote
access attempt should not abort the target process.

NETCASSINI-5630

Signed-off-by: Steve Welch <welch@hpe.com>