prov/cxi: Add ofi_index_map for multicast.
The OFI calling syntax calls fi_join_collective() for every NIC
(endpoint) participating in a collective tree, for each multicast tree
created.

The original design created a separate PTE for each collective tree and
bound the multicast address to the PTE. This is expensive in terms of
memory, and the scarcity of PTE resources puts an unacceptable cap on
the number of separate trees that can be supported.

This adds use of ofi_idm_*() functions to map multicast addresses to
multicast objects, and binds the map to the endpoint. All multicast
addresses can share the same endpoint PTE, allowing up to a million
multicast trees to be bound to the same PTE.

The COMM_KEY_RANK model used in NETSIM (cxitest) regression testing
simulates a network of multiple endpoints, and thus requires a PTE for
each endpoint in the simulated collective tree. It supports only a
single tree of limited size (32 endpoints).

An additional recv_cnt for the PTE structure is needed to pass low-level
tests that count all events that pass through the PTE, which is no
longer the same as events that pass through the multicas address.

Signed-off-by: Joe Nemeth <joseph.nemeth@hpe.com>