prov/cxi: Support FI_AV_USER_ID
SHM provider approach for FI_AV_USER_ID is taken. That is that even
though FI_AV_USER_ID is a capability, it was never defined as a primary
or secondary capability. Because of this, the CXI provider cannot
advertise this capability. Instead, the CXI provider is always set up to
support FI_AV_USER_ID and will honor the FI_AV_USER_ID flag when passed
into AV insert.

NETCASSINI-5707

Signed-off-by: Ian Ziemba <ian.ziemba@hpe.com>