prov/cxi: Serialize fi_join_collective.
NETCASSINI-5559
NETCASSINI-5577
NETCASSINI-5729

Commit 1 of 2.

fi_join_collective() must report errors if an endpoint attempts to
re-use a multicast address, or use an endpoint as hwroot for more than
one multicast tree. This information is only known to index 0 of the
collective until the final state of the join operation, which implies a
race if fi_join_collective() operations are concurrent.

The primary reason to make fi_join_collective() concurrent was to be
able to exploit the concurrency of the CURL operation. However, the FM
does not process CURL requests concurrently, for much the same kinds of
race considerations. The amount of time saved by trying to perform the
other operations concurrently is negligible.

This modification replaces the coll_ref counter -- which was only used
as a 0-or-1 flag -- with a bool join_busy, and extends its scope to the
entire join operation.

Signed-off-by: Joe Nemeth <joseph.nemeth@hpe.com>