prov/cxi: Define func to find domain cmdq
To support scalable transmit FI_AV_AUTH_KEY endpoints, command queues
must be owned by the domain. The domain is going to have to manage
access to multiple endpoints attempting to send RDMA traffic on
different VNIs. To support this, multiple transmit command queues are
required.

cxip_domain_find_cmdq() is the logic to select which transmit command
queue is used. The following is the priority:
1. Reuse command queue matching requested configuration
2. Reuse command queue which is empty
3. Allocate new command queue

Signed-off-by: Ian Ziemba <ian.ziemba@hpe.com>