prov/cxip: Fix cxip_trace problems.
Change cxip_trace_filename to cxip_trace_pathname and allow tracing to
occur in alternate directories, which is useful when the current path is
not writable by the user.

Initialization fails without initializing if no masks are selected,
preventing creation of empty files.

Early model of initializing only once at test login was flawed. This now
can be initialized, disabled, and re-initialized.

Signed-off-by: Joe Nemeth <joseph.nemeth@hpe.com>