prov/cxi: Drive TRACE() by env variables only.
Note: all CXIP_TRACE* references changed to CXIP_COLL_TRACE*
Note: all cxip_trace* references changed to cxip_coll_trace*

The TRACE() macros produce debugging traces to files that can be on a
shared file system, or local to a physical node (and could be memory
storage) for debugging collectives, which perform coordinated actions
across multiple nodes. This not only prevents implicit synchronization
of operations through shared file system waits, but also prevent
mangling of the output when using normal character buffering from
multiple sources, which is usually faster than line buffering.

This was originally put together for use with bench tests that are part
of the libfabric suite, and required initialization through function
calls within the bench tests, which makes this feature unavailable to
to external applications.

This commit refactors the TRACE() system to allow it to be entirely
configured through environment variables, and can be used with
production applications.

If the ENABLE_DEBUG flag is zero, all of the TRACE featues are removed
entirely: embedded TRACE() calls are a syntactically-robust NOOP that
does not emit code during compilation.

Otherwise, individual trace features must be activated through
environment variables, allowing different areas of code to be traced
selectively. If no trace features are selected, the trace files are not
created.

The original design also used function pointer indirection to allow all
of the trace functions to be entirely replaced. This was confusing to
maintain, and offers no real benefit.

The former cxip_coll_trace_enable() function was overloaded with multiple
purposes. This has been simplified into cxip_coll_trace_init() and
cxip_coll_trace_close(), which are automatically called during coll module
initialization, and a global cxip_coll_trace_muted flag that can be used to
temporarily mute tracing. This allows repeated reductions (for instance)
to be traced during set up, but then muted during a fast loop.

Signed-off-by: Joe Nemeth <joseph.nemeth@hpe.com>