prov/cxi: Fix fi_cntr_read/readerr to return current value
Previously fi_cntr_read/readerr would return the current
counter value in the write back buffer ( the cached value).
The Libfabric API indicates the actual count should be returned.

The routines were updated to return the actual hardware value
at the time of the call. This behavior is required for applications
for instance that want a byte count, but don't know what it should be.

The previous behavior is useful for applications that know what
the expected event count is and poll until reached. A flag was
added, FI_CXI_CNTR_CACHED, which maintains the previous semantic
and should be passed in the counter attributes flags field.

NETCASSINI-6335

Signed-off-by: Steve Welch <welch@hpe.com>