prov/cxi: Add support for 128-bit atomic CSWAP
Add FI_CSWAP support for FI_INT128 and FI_UINT128 (~6 lines); modify
atomic tests to support new types. (~90 lines)

Note: When the 128-bit int types were added to the libfabric API,
FI_DATATYPE_LAST was deprecated and could be no longer used to size
arrays or as a known invalid type. Uses of FI_DATATYPE_LAST in the
changed code as an invalid type were replaced by FI_VOID, which looks
to be stable as an invalid type. Any other uses of FI_DATATYPE_LAST
were replaced in a manner appropriate to the specific use.

NETCASSINI-986

Signed-off-by: John Byrne <john.l.byrne@hpe.com>