prov/cxi: Avoid S/W EP alt_read TX credit deadlock
When issuing a RGet from software for the alt_read
rendezvous protocol, use the same TX credit already
allocated for the done notify message. Allocating a
second credit enables a credit deadlock where the
RX queue is waiting for TX credit to issue the RGet
and the TX queue is waiting for TX credit to issue
the done notify message.

NETCASSINI-6534

Signed-off-by: Steve Welch <welch@hpe.com>