prov/cxi: Improve CP locking
In a threaded env where each thread has its own EP, each thread may be
issuing operations resulting in the changing of communication
profiles. Since communication profiles are a domain resources, this may
result in contention on the LNI CP lock. Change this lock to rw to
alleviate contention.

NETCASSINI-6750

Signed-off-by: Ian Ziemba <ian.ziemba@hpe.com>