prov/util: Define monitor unsubscribe_on_delete
Not all memory montiors have a 1-N relationship between subscribe and
the MR cache entries. Some memory monitors, such as kdreg2, have a 1-1
relationship and require unsubscribe to be called when the corresponding
MR cache entry is deleted.

To meet this requirement, unsubscribe_on_delete is defined. When true
for a memory monitor, ofi_monitor_unsubscribe() will be called on MR
cache being freed.

NETCASSINI-6959

Signed-off-by: Mike Uttormark <mike.uttormark@hpe.com>
Signed-off-by: Ian Ziemba <ian.ziemba@hpe.com>