[v2.3.x] prov/cxi: implement result seqno tracking
The resno (result sequence number) tracking was not completely
implemented, making it impossible for the leaf to determine
when it received the expected retry result from the root.
Store the resno in the reduction struct and alter the tracing
accordingly.

Signed-off-by: John Heemstra <john.heemstra@hpe.com>
(cherry picked from commit 606b7100e9d78320377e8d35bccadb41a18b9291)